# {{cookiecutter.project_slug}}/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "企业级FastAPI后端模板",
    "description": "企业级FastAPI后端模板 - 开箱即用",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/list": {
      "get": {
        "tags": [
          "用户模块"
        ],
        "summary": "查看用户列表",
        "operationId": "list_user_api_v1_users_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "每页数量",
              "default": 10,
              "title": "Page Size"
            },
            "description": "每页数量"
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "用户名称，用于搜索",
              "default": "",
              "title": "Username"
            },
            "description": "用户名称，用于搜索"
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "邮箱地址",
              "default": "",
              "title": "Email"
            },
            "description": "邮箱地址"
          },
          {
            "name": "dept_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "部门ID",
              "title": "Dept Id"
            },
            "description": "部门ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/get": {
      "get": {
        "tags": [
          "用户模块"
        ],
        "summary": "查看用户",
        "operationId": "get_user_api_v1_users_get_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "用户ID",
              "title": "User Id"
            },
            "description": "用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/create": {
      "post": {
        "tags": [
          "用户模块"
        ],
        "summary": "创建用户",
        "operationId": "create_user_api_v1_users_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/update": {
      "post": {
        "tags": [
          "用户模块"
        ],
        "summary": "更新用户",
        "operationId": "update_user_api_v1_users_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/delete": {
      "delete": {
        "tags": [
          "用户模块"
        ],
        "summary": "删除用户",
        "operationId": "delete_user_api_v1_users_delete_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "用户ID",
              "title": "User Id"
            },
            "description": "用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/reset_password": {
      "post": {
        "tags": [
          "用户模块"
        ],
        "summary": "重置密码",
        "operationId": "reset_password_api_v1_users_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_reset_password_api_v1_users_reset_password_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/api/list": {
      "get": {
        "tags": [
          "API模块"
        ],
        "summary": "查看API列表",
        "operationId": "list_api_api_v1_api_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "每页数量",
              "default": 10,
              "title": "Page Size"
            },
            "description": "每页数量"
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "API路径",
              "title": "Path"
            },
            "description": "API路径"
          },
          {
            "name": "summary",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "API简介",
              "title": "Summary"
            },
            "description": "API简介"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "API模块",
              "title": "Tags"
            },
            "description": "API模块"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api/get": {
      "get": {
        "tags": [
          "API模块"
        ],
        "summary": "查看Api",
        "operationId": "get_api_api_v1_api_get_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Api",
              "title": "Id"
            },
            "description": "Api"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api/create": {
      "post": {
        "tags": [
          "API模块"
        ],
        "summary": "创建Api",
        "operationId": "create_api_api_v1_api_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/api/update": {
      "post": {
        "tags": [
          "API模块"
        ],
        "summary": "更新Api",
        "operationId": "update_api_api_v1_api_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/api/delete": {
      "delete": {
        "tags": [
          "API模块"
        ],
        "summary": "删除Api",
        "operationId": "delete_api_api_v1_api_delete_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "api_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ApiID",
              "title": "Api Id"
            },
            "description": "ApiID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiCreate": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "API路径",
            "example": "/api/v1/user/list"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "API简介",
            "default": "",
            "example": "查看用户列表"
          },
          "method": {
            "$ref": "#/components/schemas/MethodType",
            "description": "API方法",
            "example": "GET"
          },
          "tags": {
            "type": "string",
            "title": "Tags",
            "description": "API标签",
            "example": "User"
          }
        },
        "type": "object",
        "required": [
          "path",
          "method",
          "tags"
        ],
        "title": "ApiCreate"
      },
      "ApiUpdate": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "API路径",
            "example": "/api/v1/user/list"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "API简介",
            "default": "",
            "example": "查看用户列表"
          },
          "method": {
            "$ref": "#/components/schemas/MethodType",
            "description": "API方法",
            "example": "GET"
          },
          "tags": {
            "type": "string",
            "title": "Tags",
            "description": "API标签",
            "example": "User"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "path",
          "method",
          "tags",
          "id"
        ],
        "title": "ApiUpdate"
      },
      "Body_reset_password_api_v1_users_reset_password_post": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "用户ID"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "Body_reset_password_api_v1_users_reset_password_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MethodType": {
        "type": "string",
        "enum": [
          "GET",
          "POST",
          "PUT",
          "DELETE",
          "PATCH"
        ],
        "title": "MethodType"
      },
...
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "example": "admin@qq.com"
          },
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_]+$",
            "title": "Username",
            "description": "用户名（3-20位字母数字下划线）",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "密码（至少8位，包含字母和数字）",
            "example": "AdminPass123"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "default": true
          },
          "is_superuser": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Superuser",
            "default": false
          },
          "role_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Ids",
            "default": []
          },
          "dept_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dept Id",
            "description": "部门ID",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "email",
          "username",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserUpdate": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "default": true
          },
          "is_superuser": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Superuser",
            "default": false
          },
          "role_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Ids",
            "default": []
          },
          "dept_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dept Id",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "username"
        ],
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}